Make lora as separate extensions
